Organize Windows Cairo implementation files https://bugs.webkit.org/show_bug.cgi?id=269593
Reviewed by Brent Fulgham.

Put `PLATFORM(WIN) && USE(CAIRO)` files into a `win/cairo` directory to make
it easier to determine where Cairo specific implementations live. Merge and
remove files that had no Cairo specific implementations but were named as if
they did.

* Source/WebCore/PlatformWin.cmake:
* Source/WebCore/page/win/FrameCairoWin.cpp: Removed.
* Source/WebCore/page/win/FrameWin.cpp:
* Source/WebCore/platform/graphics/win/FontCacheWin.cpp:
* Source/WebCore/platform/graphics/win/FontCustomPlatformDataWin.cpp:
* Source/WebCore/platform/graphics/win/FontWin.cpp:
* Source/WebCore/platform/graphics/win/GlyphPageTreeNodeWin.cpp: Renamed from Source/WebCore/platform/graphics/win/GlyphPageTreeNodeCairoWin.cpp.
* Source/WebCore/platform/graphics/win/GraphicsContextWin.cpp:
* Source/WebCore/platform/graphics/win/ImageAdapterWin.cpp:
* Source/WebCore/platform/graphics/win/MediaPlayerPrivateMediaFoundation.cpp:
* Source/WebCore/platform/graphics/win/SimpleFontDataCairoWin.cpp: Removed.
* Source/WebCore/platform/graphics/win/SimpleFontDataWin.cpp:
* Source/WebCore/platform/graphics/win/cairo/FontCacheWinCairo.cpp: Copied from Source/WebCore/platform/graphics/win/SimpleFontDataWin.cpp.
* Source/WebCore/platform/graphics/win/cairo/FontCustomPlatformDataWinCairo.cpp: Added.
* Source/WebCore/platform/graphics/win/cairo/FontPlatformDataWinCairo.cpp: Renamed from Source/WebCore/platform/graphics/win/FontPlatformDataCairoWin.cpp.
* Source/WebCore/platform/graphics/win/cairo/GraphicsContextWinCairo.cpp: Renamed from Source/WebCore/platform/graphics/win/GraphicsContextCairoWin.cpp.
* Source/WebCore/platform/graphics/win/cairo/ImageAdapterWinCairo.cpp: Copied from Source/WebCore/platform/graphics/win/ImageAdapterWin.cpp.
* Source/WebCore/platform/graphics/win/cairo/MediaPlayerPrivateMediaFoundationCairo.cpp: Added.
* Source/WebCore/platform/win/cairo/DragImageWinCairo.cpp: Renamed from Source/WebCore/platform/win/DragImageCairoWin.cpp.

Canonical link: https://commits.webkit.org/274920@main
