+++ Source/WebCore/platform/graphics/win/SimpleFontDataWin.cpp
@@ -46,6 +46,62 @@ void Font::platformCharWidthInit()
     initCharWidths();
 }
 
+void Font::platformInit()
+{
+    m_syntheticBoldOffset = m_platformData.syntheticBold() ? 1.0f : 0.f;
+
+    if (!m_platformData.size()) {
+        m_fontMetrics.reset();
+        m_avgCharWidth = 0;
+        m_maxCharWidth = 0;
+        return;
+    }
+
+    HWndDC dc(0);
+    SaveDC(dc);
+
+    const double metricsMultiplier = 1. / cWindowsFontScaleFactor;
+    HGDIOBJ oldFont = SelectObject(dc, m_platformData.hfont());
+
+    wchar_t faceName[LF_FACESIZE];
+    GetTextFace(dc, LF_FACESIZE, faceName);
+
+    OUTLINETEXTMETRIC metrics;
+    if (!GetOutlineTextMetrics(dc, sizeof(metrics), &metrics))
+        return;
+
+    float xHeight = metrics.otmTextMetrics.tmAscent * 0.56f; // Best guess for xHeight if no x glyph is present.
+    GLYPHMETRICS gm;
+    static const MAT2 identity = { { 0, 1 }, { 0, 0 }, { 0, 0 }, { 0, 1 } };
+    DWORD len = GetGlyphOutline(dc, 'x', GGO_METRICS, &gm, 0, 0, &identity);
+    if (len != GDI_ERROR && gm.gmptGlyphOrigin.y > 0)
+        xHeight = gm.gmptGlyphOrigin.y;
+
+    SelectObject(dc, oldFont);
+    RestoreDC(dc, -1);
+
+    // Disable antialiasing when rendering with Ahem because many tests require this.
+    if (!_wcsicmp(faceName, L"Ahem"))
+        m_allowsAntialiasing = false;
+
+    // FIXME: Needs to take OS/2 USE_TYPO_METRICS flag into account
+    // https://bugs.webkit.org/show_bug.cgi?id=199186
+    float ascent = metrics.otmTextMetrics.tmAscent * metricsMultiplier;
+    float descent = metrics.otmTextMetrics.tmDescent * metricsMultiplier;
+    float capHeight = (metrics.otmTextMetrics.tmAscent - metrics.otmTextMetrics.tmInternalLeading) * metricsMultiplier;
+    float lineGap = metrics.otmTextMetrics.tmExternalLeading * metricsMultiplier;
+
+    m_fontMetrics.setAscent(ascent);
+    m_fontMetrics.setDescent(descent);
+    m_fontMetrics.setCapHeight(capHeight);
+    m_fontMetrics.setLineGap(lineGap);
+    m_fontMetrics.setLineSpacing(lroundf(ascent) + lroundf(descent) + lroundf(lineGap));
+    m_fontMetrics.setUnitsPerEm(metrics.otmEMSquare);
+    m_fontMetrics.setXHeight(xHeight * metricsMultiplier);
+    m_avgCharWidth = metrics.otmTextMetrics.tmAveCharWidth * metricsMultiplier;
+    m_maxCharWidth = metrics.otmTextMetrics.tmMaxCharWidth * metricsMultiplier;
+}
+
 void Font::platformDestroy()
 {
     ScriptFreeCache(&m_scriptCache);
@@ -62,6 +118,32 @@ RefPtr<Font> Font::platformCreateScaledFont(const FontDescription&, float scaleF
     winfont.lfHeight = -lroundf(scaledSize * cWindowsFontScaleFactor);
     auto hfont = adoptGDIObject(::CreateFontIndirect(&winfont));
     return Font::create(FontPlatformData(WTFMove(hfont), scaledSize, m_platformData.syntheticBold(), m_platformData.syntheticOblique(), m_platformData.customPlatformData()), origin());
+}
+
+bool Font::platformSupportsCodePoint(char32_t character, std::optional<char32_t> variation) const
+{
+    return variation ? false : glyphForCharacter(character);
+}
+
+void Font::determinePitch()
+{
+    if (origin() == Origin::Remote) {
+        m_treatAsFixedPitch = false;
+        return;
+    }
+
+    // TEXTMETRICS have this. Set m_treatAsFixedPitch based off that.
+    HWndDC dc(0);
+    SaveDC(dc);
+    SelectObject(dc, m_platformData.hfont());
+
+    // Yes, this looks backwards, but the fixed pitch bit is actually set if the font
+    // is *not* fixed pitch. Unbelievable but true!
+    TEXTMETRIC tm;
+    GetTextMetrics(dc, &tm);
+    m_treatAsFixedPitch = !(tm.tmPitchAndFamily & TMPF_FIXED_PITCH);
+
+    RestoreDC(dc, -1);
 }
 
 } // namespace WebCore
